Team Bonus

+++ ServiceAutomation/ServiceAutomation.Canvas.WebApi/Models/ResponseModels/AccuralResponseModel.cs
@@ -7,7 +7,7 @@ namespace ServiceAutomation.Canvas.WebApi.Models.ResponseModels
         public Guid Id { get; set; }
         public string AccuralName { get; set; }
         public string ReferralName { get; set; }
-        public int AccuralPercent { get; set; }
+        public double AccuralPercent { get; set; }
         public string TransactionStatus { get; set; }
         public decimal InitialAmount { get; set; }
         public decimal AccuralAmount { get; set; }
